edit category and completed remaining functions

+++ src/main/java/com/dev/ecommerce/dto/OrderDTO.java
@@ -29,7 +29,7 @@ public class OrderDTO {
     private Address shippingAddress;
     private PaymentDetails paymentDetails = new PaymentDetails();
     private double totalMrpPrice;
-    private Integer totalSellingPrice;
+    private double totalSellingPrice;
     private Integer discount;
     private OrderStatus orderStatus;
     private int totalItem;
