+++ src/main/java/com/dev/ecommerce/model/SellerReport.java
@@ -17,15 +17,15 @@ public class SellerReport extends abstractEntity{
     @OneToOne
     private Seller seller;
 
-    private Long totalEarnings = 0L;
+    private Double totalEarnings = 0.0;
 
     private Long totalSales = 0L;
 
-    private Long totalRefunds = 0L;
+    private Double totalRefunds = 0.0;
 
     private Long totalTax = 0L;
 
-    private Long netEarnings = 0L;
+    private Double netEarnings = 0.0;
 
     private Integer totalOrders = 0;
